Remove ToPointerAddressNode and unwrap handles as longs directly
* VALUEs from native code are always plain longs now, never interop
  pointer objects or ValueWrappers, so the interop asPointer path of
  ToPointerAddressNode and the generic UnwrapNode accepting
  ValueWrapper-or-long are both dead.
* Rename UnwrapNativeNode to UnwrapNode and NativeToWrapperNode to
  ToWrapperNode, and make ToWrapperNode and UnwrapCArrayNode top-level
  classes in org.truffleruby.cext.
* Use them with a long argument directly in cext_unwrap,
  UnwrapCArrayNode, NativeArrayStorage, upcall V arguments,
  IDToSymbolNode, ReadCStringNode, ReadCValueNode, cext_to_wrapper,
  rb_gc_mark and rb_tr_gc_guard (their upcalls all use the L carrier).
