+++ tool/jt.rb
@@ -2947,6 +2947,7 @@ module Commands
     %r{\Asrc/yarp/},                                                    # imported from prism
     %r{\Asrc/dev\.truffleruby\.shadowed\.},                             # shadowed libraries such as joni
     %r{\Aspec/truffle/capi/ext/internal_id\.h\z},                       # imported from MRI id.h
+    %r{\Asrc/main/java/org/truffleruby/cext/(ToWrapperNode|UnwrapCArrayNode)\.java\z}, # moved out of UnwrapNode.java
   ]
 
   # The standard BSD-3-Clause license text, as in the headers of
